# docs/static/swagger.json
{
...
  "info": {
    "description": "Documenta\u00e7\u00e3o autom\u00e1tica das rotas do sistema com arquitetura DDD e SQL Puro.",
    "title": "Documenta\u00e7\u00e3o da API - Projeto g03",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users/": {
      "get": {
        "description": "Suporta pagina\u00e7\u00e3o e filtro por cargo. Acesso restrito a gestores remotos.<br/>",
        "parameters": [
          {
            "default": 1,
            "description": "N\u00famero da p\u00e1gina",
            "in": "query",
            "name": "page",
            "type": "integer"
          },
          {
            "default": 20,
            "description": "Quantidade de registros por p\u00e1gina (m\u00e1x. 100)",
            "in": "query",
            "name": "per_page",
            "type": "integer"
          },
          {
            "description": "Filtrar por cargo (ex: gestor_local, gestor_remoto)",
            "in": "query",
            "name": "role",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de usu\u00e1rios"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "403": {
            "description": "Cargo insuficiente"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Lista todos os usu\u00e1rios do sistema",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "description": "Qualquer usu\u00e1rio com token v\u00e1lido pode ver o pr\u00f3prio perfil.<br/>",
        "responses": {
          "200": {
            "description": "Dados do usu\u00e1rio logado"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "404": {
            "description": "Usu\u00e1rio n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Retorna os dados do usu\u00e1rio autenticado",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    },
    "/api/users/me/password": {
      "patch": {
        "description": "Exige a senha atual para confirmar a opera\u00e7\u00e3o.<br/>",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "new_password": {
                  "example": "nova_senha_forte456",
                  "type": "string"
                },
                "old_password": {
                  "example": "senha_antiga123",
                  "type": "string"
                }
              },
              "required": [
                "old_password",
                "new_password"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso"
          },
          "400": {
            "description": "Senha atual incorreta ou nova senha inv\u00e1lida"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Atualiza a senha do pr\u00f3prio usu\u00e1rio autenticado",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    },
    "/api/users/search": {
      "get": {
        "description": "Realiza uma busca case-insensitive pelo nome informado.<br/>",
        "parameters": [
          {
            "description": "Termo de busca (m\u00ednimo 2 caracteres)",
            "in": "query",
            "name": "q",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usu\u00e1rios encontrados"
          },
          "400": {
            "description": "Par\u00e2metro de busca inv\u00e1lido"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "403": {
            "description": "Cargo insuficiente"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Busca usu\u00e1rios pelo nome (parcial)",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    },
    "/api/users/{user_id}": {
      "delete": {
        "description": "Esta a\u00e7\u00e3o \u00e9 irrevers\u00edvel. Restrita a gestores remotos.<br/>",
        "parameters": [
          {
            "description": "UUID do usu\u00e1rio a ser removido",
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Usu\u00e1rio removido com sucesso"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "403": {
            "description": "Cargo insuficiente"
          },
          "404": {
            "description": "Usu\u00e1rio n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Remove um usu\u00e1rio do sistema permanentemente",
        "tags": [
          "Usu\u00e1rios"
        ]
      },
      "get": {
        "parameters": [
          {
            "description": "UUID do usu\u00e1rio",
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do usu\u00e1rio encontrado"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "403": {
            "description": "Cargo insuficiente"
          },
          "404": {
            "description": "Usu\u00e1rio n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Busca um usu\u00e1rio pelo seu UUID",
        "tags": [
          "Usu\u00e1rios"
        ]
      },
      "patch": {
        "description": "Apenas gestores remotos ou gestores locais podem alterar dados de outros usu\u00e1rios.<br/>",
        "parameters": [
          {
            "description": "UUID do usu\u00e1rio a ser atualizado",
            "in": "path",
            "name": "user_id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "name": {
                  "example": "Jo\u00e3o Silva Atualizado",
                  "type": "string"
                },
                "role": {
                  "example": "gestor_remoto",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usu\u00e1rio atualizado com sucesso"
          },
          "400": {
            "description": "Dados inv\u00e1lidos"
          },
          "401": {
            "description": "Token JWT ausente ou inv\u00e1lido"
          },
          "403": {
            "description": "Cargo insuficiente"
          },
          "404": {
            "description": "Usu\u00e1rio n\u00e3o encontrado"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Atualiza os dados de um usu\u00e1rio (name e/ou role)",
        "tags": [
          "Usu\u00e1rios"
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "description": "Insira o token JWT no formato: Bearer {seu_token}",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}